+++ crates/forge/src/main.rs
@@ -27,6 +27,7 @@ use std::env;
 use std::sync::Arc;
 use std::thread::available_parallelism;
 use tokio::runtime::Builder;
+use universal_sierra_compiler_api::UniversalSierraCompilerCommand;
 
 mod init;
 
@@ -302,6 +303,7 @@ fn main_execution() -> Result<bool> {
     let cli = Cli::parse();
 
     ScarbCommand::new().ensure_available()?;
+    UniversalSierraCompilerCommand::ensure_available()?;
 
     match cli.subcommand {
         ForgeSubcommand::Init { name } => {